remove TestAdminLedgerClient
